Adding Toggle trait for later use.
