+++ src/Toggle/Local.php
@@ -4,11 +4,14 @@ declare(strict_types=1);
 
 namespace FeatureToggle\Toggle;
 
+use FeatureToggle\Traits\Toggle;
 use Illuminate\Contracts\Support\Arrayable;
 use FeatureToggle\Contracts\Toggle as ToggleContract;
 
 class Local implements ToggleContract, Arrayable
 {
+    use Toggle;
+
     /**
      * @var string
      */
@@ -34,33 +37,5 @@ class Local implements ToggleContract, Arrayable
         } elseif (is_string($isActive) || is_int($isActive)) {
             $this->is_active = filter_var($isActive, FILTER_VALIDATE_BOOLEAN);
         }
-    }
-
-    /**
-     * @return string
-     */
-    public function getName(): string
-    {
-        return $this->name;
-    }
-
-    /**
-     * Check if toggle is active.
-     *
-     * @return bool
-     */
-    public function isActive(): bool
-    {
-        return $this->is_active;
-    }
-
-    /**
-     * Get the instance as an array.
-     *
-     * @return array
-     */
-    public function toArray(): array
-    {
-        return ['name' => $this->getName(), 'is_active' => $this->isActive()];
     }
 }
